improve code

+++ app/src/main/java/ir/mahoorsoft/app/cityneed/model/struct/ResponseOfServer.java
@@ -7,7 +7,7 @@ import android.util.Base64;
  */
 
 public class ResponseOfServer {
-    public int code;
+    public float code;
     public String name;
     public String apiCode;
     public String ms;
